+++ docs/next/pages/migration.tsx
@@ -93,7 +93,7 @@ export const getStaticProps: GetStaticProps = async () => {
   const pathToMdxFile = path.resolve('../../MIGRATION.md');
 
   try {
-    // 2. Read and parse versioned MDX content
+    // 2. Read and parse MDX content
     const source = await fs.readFile(pathToMdxFile);
     const {content, data} = matter(source);
